# swagger.json
{
    "swagger":"2.0",
    "info": {
        "version" : "1.0.0",
        "title" : "Ecommerce App",
        "description" : "My Ecommcerce Project API",
        "license" : {
            "name" : "MIT",
            "url":"https://opensource.org/licenses/MIT"
        }
    },
    "host" : "localhost:8000",
    "basePath" : "/",
...
    "schemes": ["http"],
    "consumes": ["application/json"],
    "produces" : ["application/json"],
    "paths" : {
...
        "/api/product/add_product":{
            "post":{
                "tags" : ["Product"],
                "summary":"Add One Product",
                "parameters":[
                    {
                        "name": "body",
                        "in" : "body",
                        "description" : "Your Username",
                        "required": true,
                        "schema" :{
                            "properties":{
                                "id": {
                                    "type": "string"
                                },
                                "name":{
                                    "type":"string"
                                },
                                "image":{
                                    "type":"string"
                                },
                                "price":{
                                    "type":"string"
                                },
                                "numberOnSale":{
                                    "type":"integer"
                                }
                            }
                        }
                    }
                ],
                "responses":{
                    "200":{
                            "message": "OK",
                            "description" : "OK",
                            "schema":{
                                "$ref" :  "#/definitions/product"
                            }
                    }
                }
            }
        },
        "/api/product/update/{id}":{
            "put":{
                "tags" : ["Product"],
                "summary":"Update Product in System",
                "parameters":[
                    {
                        "in":"path",
                        "name":"id",
                        "description":"Product ID",
                        "required": true,
                        "schema":{
                            "type":"string"
                        }
                    },
                    {
                        "name": "body",
                        "in" : "body",
                        "description" : "Your Product",
                        "required": true,
                        "schema" :{
                            "properties":{
                                "name": {
                                    "type": "string"
                                },
                                "image":{
                                    "type":"string"
                                },
                                "price":{
                                    "type":"string"
                                }
                            }
                        }
                    }

                ],
                "responses":{
                    "200":{
                            "message": "OK",
                            "description" : "OK",
                            "schema":{
                                "$ref" :  "#/definitions/product"
                            }
                    }
                }
            }
        },
...
        "/api/order/get_ord/{orderid}" :{
            "get" :{
                "tags" : ["Orders"],
                "summary" : "Get Order With ID",
                "parameters":[
                    {
                        "in":"path",
                        "name":"orderid",
                        "description":"Order ID",
                        "required": true,
                        "schema":{
                            "type":"string"
                        }
                    }],
                "responses":{
                    "200" : {
                        "message": "OK",
                        "description" : "OK",
                        "schema":{
                            "$ref" :  "#/definitions/orders"
                        }
                    }
                }
            }
        },
        "/api/order/find_order_cus/{customerid}" :{
            "get" :{
                "tags" : ["Orders"],
                "summary" : "Get User have ID",
                "parameters":[
                    {
                        "in":"path",
                        "name":"customerid",
                        "description":"Customer ID",
                        "required": true,
                        "schema":{
                            "type":"string"
                        }
                    }],
                "responses":{
                    "200" : {
                        "message": "OK",
                        "description" : "OK",
                        "schema":{
                            "$ref" :  "#/definitions/orders"
                        }
                    }
                }
            }
        },
        "/api/order/add_order":{
            "post":{
                "tags" : ["Orders"],
                "summary":"Add One Order",
                "parameters":[
                    {
                        "name": "body",
                        "in" : "body",
                        "description" : "Your Order",
                        "required": true,
                        "schema" :{
                            "properties":{
                                "orderID": {
                                    "type": "string"
                                },
                                "customerID": {
                                    "type": "string"
                                },
                                "total": {
                                    "type": "string"
                                },
                                "shippingData": {
                                    "type":"object",
                                    "properties":
                                    {
                                        "firstName":{"type":"string"},
                                        "lastName":{"type":"string"},
                                        "email":{"type":"string"},
                                        "phone" :{"type":"string"},
                                        "city":{"type":"string"},
                                        "district":{"type":"string"},
                                        "ward":{"type":"string"},
                                        "address":{"type":"string"}
                                    }
                                },
                                "date":{
                                    "type":"string"
                                },
                                "status":{
                                    "type":"string"
                                },
                                "dateSort":{
                                    "type": "string"
                                },
                                "statistic":{
                                    "type":"boolean"
                                }
                            }
                        }
                    }
                ],
                "responses":{
                    "success":{
                            "message": "OK",
                            "description" : "OK",
                            "schema":{
                                "$ref" :  "#/definitions/orders"
                            }
                    }
                }
            }
        },
        "/api/order/update_status/{ordid}":{
            "put":{
                "tags" : ["Orders"],
                "summary":"Update Order Status",
                "parameters":[
                    {
                        "in":"path",
                        "name":"ordid",
                        "description":"Order ID",
                        "required": true,
                        "schema":{
                            "type":"string"
                        }
                    },
                    {
                        "name": "body",
                        "in" : "body",
                        "description" : "Your Product",
                        "required": true,
                        "schema" :{
                            "properties":{
                                "status": {
                                    "type": "string"
                                }
                            }
                        }
                    }

                ],
                "responses":{
                    "200":{
                            "message": "OK",
                            "description" : "OK",
                            "schema":{
                                "$ref" :  "#/definitions/orders"
                            }
                    }
                }
            }
        },
        "/api/order/delete/{orderid}" :{
            "delete" :{
                "tags" : ["Orders"],
                "summary" : "Delete Order With ID",
                "parameters":[
                    {
                        "in":"path",
                        "name":"orderid",
                        "description":"Order ID",
                        "required": true,
                        "schema":{
                            "type":"string"
                        }
                    }],
                "responses":{
                    "success" : {
                        "message": "OK",
                        "description" : "Your order is successful deleted"
                    }
                }
            }
        }
    },
    "definitions" : {
...
        "product" : {
            "required" : ["id","name","price"],
            "properties" : {
                "_id" : {
                    "type" : "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "price": {
                    "type": "string"
                },
                "numberOnSale":{
                    "type": "integer"
                }
            }
        },
        "orders" : {
            "required" : ["orderID","customerID"],
            "properties" : {
                "_id" : {
                    "type" : "string"
                },
                "orderID": {
                    "type": "string"
                },
                "customerID": {
                    "type": "string"
                },
                "total": {
                    "type": "string"
                },
                "shippingData": {
                    "type":"object",
                    "properties":
                    {
                        "firstName":{"type":"string"},
                        "lastName":{"type":"string"},
                        "email":{"type":"string"},
                        "phone" :{"type":"string"},
                        "city":{"type":"string"},
                        "district":{"type":"string"},
                        "ward":{"type":"string"},
                        "address":{"type":"string"}
                    }
                },
                "date":{
                    "type":"string"
                },
                "status":{
                    "type":"string"
                },
                "dateSort":{
                    "type": "string"
                },
                "statistic":{
                    "type":"boolean"
                }
            }
        },
...
    }
}